feat(estimate-tx): support eth
PORTAL-2636
