+++ .yarn/versions/42116f2b.yml
@@ -1,0 +1,7 @@
+releases:
+  "@fluent-wallet/estimate-tx": minor
+  "@fluent-wallet/wallet_get-next-nonce": patch
+
+declined:
+  - helios-background
+  - dapp-cross-space
